pyproject.toml
[tool.poetry]
name = "discordbotleebot-2"
version = "0.1.0"
description = ""
authors = ["repl.it user <replituser@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
python-dotenv = "^0.15.0"
tinydb = "^4.4.0"
discord = "^1.0.1"
asgiref = "^3.3.1"
jmbo-event = "^0.0.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
